Move tests of goban in its dedicated file
